fix(test_tf2estimator_ray_backend.py): remove setup
